% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\name{alcohol}
\alias{alcohol}
\title{Alcohol}
\usage{
alcohol(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{}
}
\value{
CSV saved to output_folder/alcohol_*.csv
}
\description{
Alcohol
}
