% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_hospitalizations.R
\name{all_hospitalizations}
\alias{all_hospitalizations}
\title{All hospitalizations}
\usage{
all_hospitalizations(output_folder)
}
\value{
output_folder/all_hospitalizations_*.csv
}
\description{
All hospitalizations
}
