% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiety.R
\name{anxiety}
\alias{anxiety}
\title{Anxiety}
\usage{
anxiety(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{dataset}{a string returned by Sys.getenv("WORKSPACE_CDR"). Can also use another dataset, but this is not recommended.}

\item{output_folder}{the folder to write the output to. Must be preceded by the workspace bucket location given by Sys.getenv("WORKSPACE_BUCKET").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/anxiety.csv
}
\description{
Anxiety
}
\details{
At least 1 ICD code

ICD9: "300.00","300.01","300.02"

ICD10: "F41.9","F41.0","F41.1"
}
