% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_bmi.R
\name{closest_bmi}
\alias{closest_bmi}
\title{Closest BMI}
\usage{
closest_bmi(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_bmi.csv
}
\description{
Closest BMI
}
