% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_bp.R
\name{closest_bp}
\alias{closest_bp}
\title{Closest blood pressure}
\usage{
closest_bp(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  cohort = NULL
)
}
\value{
output_folder/closest_bp.csv
}
\description{
Closest blood pressure
}
