% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_fasting_glucose.R
\name{closest_fasting_glucose}
\alias{closest_fasting_glucose}
\title{Closest fasting glucose}
\usage{
closest_fasting_glucose(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_fasting_glucose.csv
}
\description{
Closest fasting glucose
}
