% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_glucose.R
\name{closest_glucose}
\alias{closest_glucose}
\title{Closest glucose}
\usage{
closest_glucose(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_glucose.csv
}
\description{
Closest glucose
}
