% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_hdl.R
\name{closest_hdl}
\alias{closest_hdl}
\title{Closest HDL}
\usage{
closest_hdl(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_hdl.csv
}
\description{
Closest HDL
}
