% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_height.R
\name{closest_height}
\alias{closest_height}
\title{Closest height}
\usage{
closest_height(
  dataset,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  output_folder
)
}
\value{
output_folder/closest_height.csv
}
\description{
Closest height
}
