% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_hgba1c.R
\name{closest_hgba1c}
\alias{closest_hgba1c}
\title{Closest HgbA1c}
\usage{
closest_hgba1c(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_hgba1c.csv
}
\description{
Closest HgbA1c
}
