% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_ldl.R
\name{closest_ldl}
\alias{closest_ldl}
\title{Closest LDL}
\usage{
closest_ldl(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_ldl.csv
}
\description{
Closest LDL
}
