% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_pro_bnp.R
\name{closest_pro_bnp}
\alias{closest_pro_bnp}
\title{Closest Pro-BNP}
\usage{
closest_pro_bnp(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_pro_bnp.csv
}
\description{
Closest Pro-BNP
}
