% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_trigs.R
\name{closest_trigs}
\alias{closest_trigs}
\title{Closest Triglycerides}
\usage{
closest_trigs(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_trigs.csv
}
\description{
Closest Triglycerides
}
