% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_weight.R
\name{closest_weight}
\alias{closest_weight}
\title{Closest weight}
\usage{
closest_weight(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_weight.csv
}
\description{
Closest weight
}
