% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dementia.R
\name{dementia}
\alias{dementia}
\title{Dementia}
\usage{
dementia(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/dementia.csv
}
\description{
Dementia
}
