% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{Demographics}
\usage{
demographics(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/demographics_*.csv
}
\description{
Demographics
}
