% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\name{diabetes}
\alias{diabetes}
\title{Diabetes}
\usage{
diabetes(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/diabetes.csv
}
\description{
Diabetes
}
