% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display_anova}
\alias{display_anova}
\title{Display ANOVA}
\usage{
display_anova(aov, width = 150)
}
\arguments{
\item{width}{the width of the table in pixels}

\item{s}{a model returned by rms}
}
\description{
Display ANOVA
}
