% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display_df}
\alias{display_df}
\title{Display a dataframe}
\usage{
display_df(df, width = 150, digits = 3)
}
\arguments{
\item{df}{a dataframe}

\item{width}{the width of the table in pixels}

\item{digits}{the number of digits to round to}
}
\description{
Display a dataframe
}
