% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display_summary}
\alias{display_summary}
\title{Display summary}
\usage{
display_summary(s, width = 150)
}
\arguments{
\item{s}{a model returned by rms}

\item{width}{the width of the table in pixels}
}
\description{
Display summary
}
