% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fitbit_cox.R
\name{format_clust_cox}
\alias{format_clust_cox}
\title{Format clustering data for time varying Cox model.}
\usage{
format_clust_cox(clust_dat, dx, last_medical_encounter, verbose = TRUE)
}
\arguments{
\item{dx}{data.table with columns: person_id, dx_entry_date, dx_status}

\item{last_medical_encounter}{data.table with columns: person_id, last_medical_encounter_entry_date}

\item{fitbit}{data.table with columns: person_id, daily_class (can range from 1 to 5).}
}
\value{
A data.table
}
\description{
Format clustering data for time varying Cox model.
}
