% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_failure.R
\name{heart_failure}
\alias{heart_failure}
\title{Heart failure}
\usage{
heart_failure(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/heart_failure.csv
}
\description{
Heart failure
}
