% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertension.R
\name{hypertension}
\alias{hypertension}
\title{Hypertension}
\usage{
hypertension(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/hypertension.csv
}
\description{
Hypertension
}
