% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_fitbit_covariates.R
\name{impute_fitbit_covariates}
\alias{impute_fitbit_covariates}
\title{Impute covariates for time-varying fitbit dataset}
\usage{
impute_fitbit_covariates(fitbit_formatted, covariates)
}
\arguments{
\item{fitbit_formatted}{the data.table output from format_fitbit_cox}

\item{merged_covariates}{a data.table with a "person_id" column and any other covariates. Must have same person_ids as fitbit_formatted.}
}
\value{
a mids object object
}
\description{
Impute covariates for time-varying fitbit dataset
}
