% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_long_data.R
\name{impute_long_data}
\alias{impute_long_data}
\title{Impute covariates for time-varying dataset}
\usage{
impute_long_data(dat_long, covariates)
}
\arguments{
\item{dat_long}{a data.table representing a time-varying dataset}

\item{merged_covariates}{a data.table with a "person_id" column and any other covariates. Must have same person_ids as dat_long.}
}
\value{
a mids object object
}
\description{
Impute covariates for time-varying dataset
}
