% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstitial_lung_disease.R
\name{interstitial_lung_disease}
\alias{interstitial_lung_disease}
\title{Interstitial lung disease}
\usage{
interstitial_lung_disease(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/interstitial_lung_disease.csv
}
\description{
Interstitial lung disease
}
