% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_medical_encounter.R
\name{last_medical_encounter}
\alias{last_medical_encounter}
\title{Last Medical Encounter}
\usage{
last_medical_encounter(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/last_medical_encounter_*.csv
}
\description{
Last Medical Encounter
}
