% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_diuretics.R
\name{loop_diuretics}
\alias{loop_diuretics}
\title{Loop diuretics}
\usage{
loop_diuretics(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/loop_diuretics.csv
}
\description{
Loop diuretics
}
