% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/major_depressive_disorder.R
\name{major_depressive_disorder}
\alias{major_depressive_disorder}
\title{Major depressive disorder}
\usage{
major_depressive_disorder(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/major_depressive_disorder.csv
}
\description{
Major depressive disorder
}
