% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_components.R
\name{med_query_min_date}
\alias{med_query_min_date}
\title{Medication query returning first date}
\usage{
med_query_min_date(dataset, meds, page_size = NULL)
}
\description{
Medication query returning first date
}
