% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraine.R
\name{migraine}
\alias{migraine}
\title{Migraine}
\usage{
migraine(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/migraine.csv
}
\description{
Migraine
}
