% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineralocorticoid_antagonists.R
\name{mineralocorticoid_antagonists}
\alias{mineralocorticoid_antagonists}
\title{Mineralocorticoid Antagonists}
\usage{
mineralocorticoid_antagonists(
  dataset,
  output_path,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/mineralocorticoid_antagonists.csv
}
\description{
Mineralocorticoid Antagonists
}
