% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\name{obesity}
\alias{obesity}
\title{Obesity}
\usage{
obesity(
  dataset,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  output_folder
)
}
\value{
output_folder/obesity.csv
}
\description{
Obesity
}
