% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity_bmi.R
\name{obesity_bmi}
\alias{obesity_bmi}
\title{Obesity using BMI}
\usage{
obesity_bmi(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/obesity_bmi.csv
}
\description{
Obesity using BMI
}
