% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdann_sleep.R
\name{sdann_sleep}
\alias{sdann_sleep}
\title{SDANN while asleep: Standard deviation of the average NN intervals while asleep}
\usage{
sdann_sleep(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{dataset}{a string returned by Sys.getenv("WORKSPACE_CDR"). Can also use another dataset, but this is not recommended.}

\item{output_folder}{the folder to write the output to. Must be preceded by the workspace bucket location given by Sys.getenv("WORKSPACE_BUCKET").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdann_sleep.csv
}
\description{
SDANN while asleep: Standard deviation of the average NN intervals while asleep
}
\details{
Heart rate minute level data is extracted and five-minute averages are calculated for consecutive five-minute intervals.
The average HR is used to calculate the average RR duration of each five-minute interval:
Average RR = 6000 / mean(HR)
Subsequently, the standard deviation of all the five-minute RR intervals is calculated, yielding the SDANN value (in ms).
HR values are relegated to those while sleeping.
}
