% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglt2_inhibitors.R
\name{sglt2_inhibitors}
\alias{sglt2_inhibitors}
\title{SGLT2 Inhibitors}
\usage{
sglt2_inhibitors(
  dataset,
  output_path,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/sglt2_inhibitors.csv
}
\description{
SGLT2 Inhibitors
}
