% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_apnea.R
\name{sleep_apnea}
\alias{sleep_apnea}
\title{Sleep Apnea}
\usage{
sleep_apnea(
  dataset,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  output_folder
)
}
\value{
output_folder/sleep_apnea.csv
}
\description{
Sleep Apnea
}
