% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{smoking}
\alias{smoking}
\title{Smoking}
\usage{
smoking(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/smoking_*.csv
}
\description{
Smoking
}
