% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke.R
\name{stroke}
\alias{stroke}
\title{Stroke}
\usage{
stroke(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{dataset}{a string returned by Sys.getenv("WORKSPACE_CDR"). Can also use another dataset, but this is not recommended.}

\item{output_folder}{the folder to write the output to. Must be preceded by the workspace bucket location given by Sys.getenv("WORKSPACE_BUCKET").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/stroke.csv
}
\description{
Stroke
}
\details{
At lest 1 ICD code:

ICD9: "433.01","433.11","433.21","433.31","433.81","433.91","434.01","434.11","434.91"

ICD10: "I63","I63.\%"
}
