% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top1_steps.R
\name{top1_steps}
\alias{top1_steps}
\title{Top 1 Steps}
\usage{
top1_steps(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/top1_steps.csv
}
\description{
Top 1 Steps
}
