% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top20_steps.R
\name{top20_steps}
\alias{top20_steps}
\title{Top 20 Steps}
\usage{
top20_steps(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/top20_steps.csv
}
\description{
Top 20 Steps
}
