% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5_steps.R
\name{top5_steps}
\alias{top5_steps}
\title{Top 5 Steps}
\usage{
top5_steps(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/top5_steps.csv
}
\description{
Top 5 Steps
}
