% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_time.R
\name{wear_time}
\alias{wear_time}
\title{Fitbit wear time}
\usage{
wear_time(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/wear_time_*.csv
}
\description{
Fitbit wear time
}
