% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_hospitalizations.R
\name{all_hospitalizations}
\alias{all_hospitalizations}
\title{All hospitalizations}
\usage{
all_hospitalizations(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/all_hospitalizations.csv
}
\description{
All hospitalizations
}
