% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_blockers.R
\name{beta_blockers}
\alias{beta_blockers}
\title{Beta blockers}
\usage{
beta_blockers(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
CSV saved to output_folder/beta_blockers.csv
}
\description{
Beta blockers
}
