% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_bnp.R
\name{closest_bnp}
\alias{closest_bnp}
\title{Closest BNP}
\usage{
closest_bnp(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/closest_bnp.csv
}
\description{
Closest BNP
}
