% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_sbp.R
\name{closest_sbp}
\alias{closest_sbp}
\title{Closest systolic blood pressure}
\usage{
closest_sbp(
  dataset,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  output_folder
)
}
\value{
output_folder/closest_sbp.csv
}
\description{
Closest systolic blood pressure
}
