% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{Demographics}
\usage{
demographics(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{dataset}{a string returned by Sys.getenv("WORKSPACE_CDR"). Can also use another dataset, but this is not recommended.}

\item{output_folder}{the folder to write the output to. Must be preceded by the workspace bucket location given by Sys.getenv("WORKSPACE_BUCKET").}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. The dataset will be merged with demographics.}

\item{before}{serves no function}

\item{after}{serves no function}
}
\value{
output_folder/demographics_*.csv
}
\description{
Demographics
}
