% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download_data}
\alias{download_data}
\title{Download data given query}
\usage{
download_data(query)
}
\arguments{
\item{a}{SQL query string}
}
\value{
a data.table
}
\description{
Download data given query
}
