% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlipidemia.R
\name{hyperlipidemia}
\alias{hyperlipidemia}
\title{Hyperlipidemia}
\usage{
hyperlipidemia(
  dataset,
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/hyperlipidemia.csv
}
\description{
Hyperlipidemia
}
