% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_medical_encounter.R
\name{last_medical_encounter}
\alias{last_medical_encounter}
\title{Last Medical Encounter}
\usage{
last_medical_encounter(output_folder)
}
\value{
output_folder/last_medical_encounter_*.csv
}
\description{
Last Medical Encounter
}
