% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pneumonia.R
\name{pneumonia}
\alias{pneumonia}
\title{Pneumonia}
\usage{
pneumonia(
  dataset,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  output_folder
)
}
\value{
output_folder/asthma.csv
}
\description{
Pneumonia
}
